refactor(backend): split test_views.py into domain test modules
Mirror the farm domain packages in the API test layout: the 2200-line
test_views.py becomes test_suppliers_api.py, test_structure_api.py,
test_planning_api.py, test_cultures_api.py and test_notes_api.py, with
the shared authenticated-project fixture extracted into
farm/tests/api_base.py (ProjectApiTestCase). All 109 tests moved
verbatim; total backend test count unchanged at 382.
