+++ backend/farm/tests/test_cultures_api.py
@@ -1,354 +1,24 @@
-from unittest.mock import patch
-from django.contrib.auth import get_user_model
-from django.core.files.uploadedfile import SimpleUploadedFile
+"""API tests for cultures, supplier data, and seed demand."""
+
 from datetime import date
 
 from rest_framework import status
 from rest_framework.test import APITestCase as DRFAPITestCase
 
-from farm.models import Bed, BedLayout, Culture, CultureSupplierData, Field, FieldLayout, Location, PlantingPlan, Project, ProjectMembership, PublicCulture, Supplier, NoteAttachment, SeedPackage
-
-User = get_user_model()
-
-
-class ApiEndpointsTest(DRFAPITestCase):
-    def setUp(self):
-        self.user = User.objects.create_user(username='testuser', email='test@example.com', password='testpass', is_active=True)
-        self.project = Project.objects.create(name='Test Project', slug='test-project')
-        ProjectMembership.objects.create(user=self.user, project=self.project, role='admin')
-        self.client.force_authenticate(user=self.user)
-        self.client.defaults['HTTP_X_PROJECT_ID'] = str(self.project.id)
-        self.location = Location.objects.create(name="API Test Location", project=self.project)
-        self.field = Field.objects.create(name="API Test Field", location=self.location, project=self.project)
-        self.bed = Bed.objects.create(
-            name="API Test Bed",
-            field=self.field,
-            area_sqm=20.0,  # Total area: 20 sqm
-            project=self.project,
-        )
-        self.culture = Culture.objects.create(
-            name="API Test Culture",
-            growth_duration_days=7,
-            harvest_duration_days=2,
-            project=self.project,
-        )
-        self.supplier = Supplier.objects.create(name="Test Supplier Co.", homepage_url='https://test-supplier-co..example', project=self.project)
-
-    def test_supplier_list(self):
-        """Test listing suppliers"""
-        response = self.client.get('/openfarmplanner/api/suppliers/')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(response.data['results']), 1)
-        self.assertEqual(response.data['results'][0]['name'], "Test Supplier Co.")
-
-    def test_supplier_list_with_search(self):
-        """Test searching suppliers"""
-        Supplier.objects.create(name="Another Supplier", homepage_url='https://another-supplier.example', project=self.project)
-        response = self.client.get('/openfarmplanner/api/suppliers/?q=Test Supplier')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertGreaterEqual(len(response.data['results']), 1)
-        # At least one result should be "Test Supplier Co."
-        supplier_names = [s['name'] for s in response.data['results']]
-        self.assertIn("Test Supplier Co.", supplier_names)
-
-    def test_supplier_create_new(self):
-        """Test creating a new supplier"""
-        data = {'name': 'New Supplier Inc.', 'homepage_url': 'https://new-supplier.example'}
-        response = self.client.post('/openfarmplanner/api/suppliers/', data)
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        self.assertEqual(Supplier.objects.count(), 2)
-        self.assertEqual(response.data['name'], 'New Supplier Inc.')
-
-    def test_supplier_create_existing(self):
-        """Test creating supplier with exact duplicate name returns a field error."""
-        data = {'name': 'Test Supplier Co.', 'homepage_url': 'https://test-supplier.example'}
-        response = self.client.post('/openfarmplanner/api/suppliers/', data)
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertEqual(Supplier.objects.count(), 1)
-        self.assertIn('name', response.data)
-        self.assertEqual(str(response.data['name'][0]), 'Ein Lieferant mit diesem Namen existiert bereits.')
-
-    def test_supplier_create_normalized_match(self):
-        """Test creating supplier with normalized match returns a field error."""
-        data = {'name': '  TEST SUPPLIER co.  ', 'homepage_url': 'https://test-supplier.example'}
-        response = self.client.post('/openfarmplanner/api/suppliers/', data)
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertEqual(Supplier.objects.count(), 1)
-        self.assertIn('name', response.data)
-
-    def test_supplier_create_allows_same_normalized_name_in_different_projects(self):
-        """Test creating the same supplier name in two projects."""
-        other_project = Project.objects.create(name='Other Project', slug='other-project')
-        ProjectMembership.objects.create(user=self.user, project=other_project, role='admin')
-        response = self.client.post(
-            '/openfarmplanner/api/suppliers/',
-            {'name': self.supplier.name, 'homepage_url': 'https://other-supplier.example'},
-            HTTP_X_PROJECT_ID=str(other_project.id),
-        )
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        self.assertEqual(response.data['name'], self.supplier.name)
-        self.assertEqual(Supplier.objects.filter(name_normalized=self.supplier.name_normalized).count(), 2)
-
-    def test_supplier_create_rejects_trimmed_case_duplicate(self):
-        """Test creating normalized duplicates with whitespace and casing differences."""
-        Supplier.objects.create(name='Reinsaat', homepage_url='https://reinsaat.example', project=self.project)
-        response = self.client.post(
-            '/openfarmplanner/api/suppliers/',
-            {'name': ' reinsaat ', 'homepage_url': 'https://reinsaat-two.example'},
-        )
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertIn('name', response.data)
-
-    def test_supplier_create_allows_empty_homepage_url(self):
-        """Test creating supplier without a website."""
-        response = self.client.post('/openfarmplanner/api/suppliers/', {'name': 'Supplier Without Website', 'homepage_url': ''})
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        self.assertEqual(response.data['homepage_url'], '')
-
-    def test_supplier_create_allows_full_homepage_url(self):
-        """Test creating supplier with a full website URL."""
-        response = self.client.post(
-            '/openfarmplanner/api/suppliers/',
-            {'name': 'Supplier With Website', 'homepage_url': 'https://supplier.example/path'},
-        )
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        self.assertEqual(response.data['homepage_url'], 'https://supplier.example/path')
-
-    def test_supplier_update_not_limited_by_list_slice(self):
-        """Supplier updates must work even if list endpoint shows only first 20 rows."""
-        for idx in range(30):
-            Supplier.objects.create(
-                name=f"A Supplier {idx:02d}",
-                homepage_url=f"https://a-supplier-{idx:02d}.example",
-                project=self.project,
-            )
-
-        target = Supplier.objects.create(name='ZZZ Supplier', homepage_url='https://zzz.example', project=self.project)
-
-        response = self.client.put(
-            f'/openfarmplanner/api/suppliers/{target.id}/',
-            {
-                'name': 'ZZZ Supplier Updated',
-                'homepage_url': 'https://zzz-updated.example',
-            },
-            format='json',
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data['name'], 'ZZZ Supplier Updated')
-
-    def test_supplier_update_with_invalid_allowed_domain_returns_400(self):
-        """Invalid allowed_domains input should be validated, not crash with 500."""
-        response = self.client.put(
-            f'/openfarmplanner/api/suppliers/{self.supplier.id}/',
-            {
-                'name': self.supplier.name,
-                'homepage_url': 'https://lieferando.example',
-                'allowed_domains': ['lieferando.example', 'www.lieferando.example', 'sdfasdad'],
-            },
-            format='json',
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertIn('allowed_domains', response.data)
-
-    def test_supplier_create_with_invalid_domain_returns_400(self):
-        """Test creating supplier with invalid allowed_domains returns 400."""
-        data = {
-            'name': 'Invalid Domain Supplier',
-            'homepage_url': 'https://example.com',
-            'allowed_domains': ['example.com', 'invalid domain', 'https://not-allowed.com']
-        }
-        response = self.client.post('/openfarmplanner/api/suppliers/', data, format='json')
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertIn('allowed_domains', response.data)
-
-    def test_supplier_create_with_invalid_url_returns_400(self):
-        """Test creating supplier with invalid homepage_url returns 400."""
-        invalid_urls = ['invalid-url', 'htp://broken.com', 'just-text']
-        for invalid_url in invalid_urls:
-            data = {
-                'name': f'Test Supplier {invalid_url}',
-                'homepage_url': invalid_url,
-            }
-            response = self.client.post('/openfarmplanner/api/suppliers/', data, format='json')
-            self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST, f'Should reject invalid URL: {invalid_url}')
-            self.assertIn('homepage_url', response.data, f'Should have homepage_url error for: {invalid_url}')
-
-    def test_supplier_create_with_domain_only_normalizes_to_https(self):
-        """Test creating supplier with domain-only URL prepends https://."""
-        test_cases = [
-            ('example.com', 'https://example.com'),
-            ('www.example.com', 'https://www.example.com'),
-            ('subdomain.example.co.uk', 'https://subdomain.example.co.uk'),
-        ]
-        
-        for input_url, expected_url in test_cases:
-            data = {
-                'name': f'Test Supplier for {input_url}',
-                'homepage_url': input_url,
-            }
-            response = self.client.post('/openfarmplanner/api/suppliers/', data, format='json')
-            self.assertEqual(response.status_code, status.HTTP_201_CREATED, f'Should accept domain-only URL: {input_url}')
-            self.assertEqual(response.data['homepage_url'], expected_url, f'Should normalize {input_url} to {expected_url}')
-
-    def test_media_upload_rejects_non_image_file(self):
-        upload = SimpleUploadedFile('payload.txt', b'not-an-image', content_type='text/plain')
-        response = self.client.post(
-            '/openfarmplanner/api/media-files/upload/',
-            {'file': upload},
-            format='multipart',
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertIn('file', response.data)
-
-    def test_location_create_and_update_with_agronomic_fields(self):
-        create_response = self.client.post(
-            '/openfarmplanner/api/locations/',
-            {
-                'name': 'Südfläche',
-                'address': 'Dorfstraße 1',
-                'description': 'Acker hinter Hof',
-                'soil_type': 'loam',
-                'exposure': 'south',
-                'latitude': 48.1234,
-                'longitude': 16.4321,
-            },
-            format='json',
-        )
-        self.assertEqual(create_response.status_code, status.HTTP_201_CREATED)
-        self.assertEqual(create_response.data['soil_type'], 'loam')
-        self.assertEqual(create_response.data['exposure'], 'south')
-
-        location_id = create_response.data['id']
-        update_response = self.client.patch(
-            f'/openfarmplanner/api/locations/{location_id}/',
-            {'soil_type': None, 'exposure': None, 'description': ''},
-            format='json',
-        )
-        self.assertEqual(update_response.status_code, status.HTTP_200_OK)
-        self.assertIsNone(update_response.data['soil_type'])
-        self.assertIsNone(update_response.data['exposure'])
+from farm.models import (
+    Culture,
+    CultureSupplierData,
+    PlantingPlan,
+    Project,
+    ProjectMembership,
+    PublicCulture,
+    SeedPackage,
+    Supplier,
+)
+from farm.tests.api_base import ProjectApiTestCase, User
 
 
-
-    def test_field_update_with_length_and_width_overwrites_area(self):
-        response = self.client.put(
-            f'/openfarmplanner/api/fields/{self.field.id}/',
-            {
-                'name': self.field.name,
-                'location': self.location.id,
-                'area_sqm': 200.0,
-                'length_m': 30.0,
-                'width_m': 4.0,
-                'notes': '',
-                'project': self.project.id,
-            },
-            format='json',
-        )
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.field.refresh_from_db()
-        self.assertEqual(float(self.field.area_sqm), 120.0)
-        self.assertEqual(self.field.length_m, 30.0)
-        self.assertEqual(self.field.width_m, 4.0)
-
-    def test_field_update_with_single_dimension_keeps_area(self):
-        response = self.client.put(
-            f'/openfarmplanner/api/fields/{self.field.id}/',
-            {
-                'name': self.field.name,
-                'location': self.location.id,
-                'area_sqm': 200.0,
-                'length_m': 25.0,
-                'width_m': None,
-                'notes': '',
-                'project': self.project.id,
-            },
-            format='json',
-        )
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.field.refresh_from_db()
-        self.assertEqual(float(self.field.area_sqm), 200.0)
-        self.assertEqual(self.field.length_m, 25.0)
-        self.assertIsNone(self.field.width_m)
-
-    def test_field_create_rejects_location_from_other_project(self):
-        other_project = Project.objects.create(name='Other project', slug='other-project')
-        foreign_location = Location.objects.create(name='Foreign location', project=other_project)
-
-        response = self.client.post(
-            '/openfarmplanner/api/fields/',
-            {
-                'name': 'Cross-project field',
-                'location': foreign_location.id,
-                'area_sqm': 10.0,
-            },
-            format='json',
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertIn('location', response.data)
-
-    def test_planting_plan_create_rejects_bed_from_other_project(self):
-        other_project = Project.objects.create(name='Other project 2', slug='other-project-2')
-        other_location = Location.objects.create(name='Other location', project=other_project)
-        other_field = Field.objects.create(name='Other field', location=other_location, project=other_project)
-        other_bed = Bed.objects.create(name='Other bed', field=other_field, area_sqm=5.0, project=other_project)
-
-        response = self.client.post(
-            '/openfarmplanner/api/planting-plans/',
-            {
-                'culture': self.culture.id,
-                'bed': other_bed.id,
-                'planting_date': date(2026, 3, 1).isoformat(),
-                'area_usage_sqm': 1.0,
-            },
-            format='json',
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertIn('bed', response.data)
-
-    def test_bed_update_with_length_and_width_overwrites_area(self):
-        response = self.client.put(
-            f'/openfarmplanner/api/beds/{self.bed.id}/',
-            {
-                'name': self.bed.name,
-                'field': self.field.id,
-                'area_sqm': 20.0,
-                'length_m': 5.0,
-                'width_m': 3.0,
-                'notes': '',
-                'project': self.project.id,
-            },
-            format='json',
-        )
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.bed.refresh_from_db()
-        self.assertEqual(float(self.bed.area_sqm), 15.0)
-        self.assertEqual(self.bed.length_m, 5.0)
-        self.assertEqual(self.bed.width_m, 3.0)
-
-    def test_bed_update_with_single_dimension_keeps_area(self):
-        response = self.client.put(
-            f'/openfarmplanner/api/beds/{self.bed.id}/',
-            {
-                'name': self.bed.name,
-                'field': self.field.id,
-                'area_sqm': 20.0,
-                'length_m': 7.0,
-                'width_m': None,
-                'notes': '',
-                'project': self.project.id,
-            },
-            format='json',
-        )
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.bed.refresh_from_db()
-        self.assertEqual(float(self.bed.area_sqm), 20.0)
-        self.assertEqual(self.bed.length_m, 7.0)
-        self.assertIsNone(self.bed.width_m)
+class CultureApiTest(ProjectApiTestCase):
     def test_culture_with_supplier(self):
         """Test creating culture with supplier"""
         data = {
@@ -502,7 +172,7 @@ class ApiEndpointsTest(DRFAPITestCase):
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertTrue(response.data['exists'])
         self.assertEqual(response.data['culture']['name'], 'Tomate')
-    
+
     def test_culture_create_with_all_fields(self):
         """Test creating culture with all manual planning fields"""
         data = {
@@ -532,7 +202,7 @@ class ApiEndpointsTest(DRFAPITestCase):
         self.assertEqual(response.data['nutrient_demand'], 'high')
         self.assertEqual(response.data['thousand_kernel_weight_g'], 472.02)
         self.assertEqual(response.data['display_color'], '#FF5733')
-    
+
     def test_culture_create_without_durations(self):
         """Test that creating culture without durations is allowed"""
         data = {
@@ -545,7 +215,7 @@ class ApiEndpointsTest(DRFAPITestCase):
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
         self.assertIsNone(response.data['growth_duration_days'])
         self.assertIsNone(response.data['harvest_duration_days'])
-    
+
     def test_culture_create_invalid_display_color(self):
         """Test that invalid display color format is rejected"""
         data = {
@@ -559,7 +229,7 @@ class ApiEndpointsTest(DRFAPITestCase):
         response = self.client.post('/openfarmplanner/api/cultures/', data)
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
         self.assertIn('display_color', response.data)
-    
+
     def test_culture_update(self):
         """Test updating a culture"""
         data = {
@@ -602,7 +272,7 @@ class ApiEndpointsTest(DRFAPITestCase):
         self.assertEqual(response.data['thousand_kernel_weight_g'], 3.8)
         self.culture.refresh_from_db()
         self.assertEqual(float(self.culture.thousand_kernel_weight_g), 3.8)
-    
+
     def test_culture_update_with_seed_packages_payload_from_get(self):
         """PUT with seed package objects (including id/culture) should stay valid."""
         supplier = Supplier.objects.create(name='Seed Supplier', homepage_url='https://seed-supplier.example', project=self.project)
@@ -883,34 +553,6 @@ class ApiEndpointsTest(DRFAPITestCase):
         self.assertEqual(row['seed_packages'], [{'size_value': 25.0, 'size_unit': 'g'}])
         self.assertEqual(row['package_suggestion']['pack_count'], 1)
 
-
-    def test_bed_list(self):
-        response = self.client.get('/openfarmplanner/api/beds/')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(response.data['results']), 1)
-
-    def test_planting_plan_create(self):
-        data = {
-            'culture': self.culture.id,
-            'bed': self.bed.id,
-            'planting_date': '2024-03-01',
-            'quantity': 50
-        }
-        response = self.client.post('/openfarmplanner/api/planting-plans/', data)
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        self.assertIn('harvest_date', response.data)
-    
-    def test_field_create_with_valid_area(self):
-        data = {
-            'name': 'Valid Field',
-            'location': self.location.id,
-            'area_sqm': 500.50,
-            'project': self.project.id,
-        }
-        response = self.client.post('/openfarmplanner/api/fields/', data)
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        self.assertEqual(Field.objects.count(), 2)
-    
     def test_culture_supplier_product_url_domain_mismatch_rejected(self):
         data = {
             'name': 'Culture with URL mismatch',
@@ -925,182 +567,6 @@ class ApiEndpointsTest(DRFAPITestCase):
         response = self.client.post('/openfarmplanner/api/cultures/', data, format='json')
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
         self.assertIn('supplier_product_url', response.data)
-
-    def test_field_create_with_invalid_area_too_small(self):
-        data = {
-            'name': 'Too Small Field',
-            'location': self.location.id,
-            'area_sqm': 0.001,
-            'project': self.project.id,
-        }
-        response = self.client.post('/openfarmplanner/api/fields/', data)
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertIn('area_sqm', response.data)
-    
-    def test_field_create_with_invalid_area_too_large(self):
-        data = {
-            'name': 'Too Large Field',
-            'location': self.location.id,
-            'area_sqm': 2000000,  # Greater than MAX_AREA_SQM (1,000,000)
-            'project': self.project.id,
-        }
-        response = self.client.post('/openfarmplanner/api/fields/', data)
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertIn('area_sqm', response.data)
-    
-    def test_bed_create_with_valid_area(self):
-        data = {
-            'name': 'Valid Bed',
-            'field': self.field.id,
-            'area_sqm': 50.2,
-            'project': self.project.id,
-        }
-        response = self.client.post('/openfarmplanner/api/beds/', data)
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        self.assertEqual(Bed.objects.count(), 2)
-    
-    def test_bed_create_with_invalid_area_too_small(self):
-        data = {
-            'name': 'Too Small Bed',
-            'field': self.field.id,
-            'area_sqm': 0.001,
-            'project': self.project.id,
-        }
-        response = self.client.post('/openfarmplanner/api/beds/', data)
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertIn('area_sqm', response.data)
-    
-    def test_bed_create_with_invalid_area_too_large(self):
-        data = {
-            'name': 'Too Large Bed',
-            'field': self.field.id,
-            'area_sqm': 20000,
-            'project': self.project.id,
-        }
-        response = self.client.post('/openfarmplanner/api/beds/', data)
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertIn('area_sqm', response.data)
-
-    def test_planting_plan_area_validation_success(self):
-        """Test API allows a single planting plan within bed capacity."""
-        data = {
-            'culture': self.culture.id,
-            'bed': self.bed.id,
-            'planting_date': '2024-03-01',
-            'area_usage_sqm': 15.0
-        }
-        response = self.client.post('/openfarmplanner/api/planting-plans/', data)
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        self.assertEqual(float(response.data['area_usage_sqm']), 15.0)
-
-
-
-    def test_planting_plan_area_validation_non_overlapping_plans_allowed(self):
-        """Test API allows non-overlapping plans even when their total sum is above bed capacity."""
-        response1 = self.client.post('/openfarmplanner/api/planting-plans/', {
-            'culture': self.culture.id,
-            'bed': self.bed.id,
-            'planting_date': '2024-03-01',
-            'area_usage_sqm': 20.0,
-        })
-        self.assertEqual(response1.status_code, status.HTTP_201_CREATED)
-
-        response2 = self.client.post('/openfarmplanner/api/planting-plans/', {
-            'culture': self.culture.id,
-            'bed': self.bed.id,
-            'planting_date': '2024-04-01',
-            'area_usage_sqm': 20.0,
-        })
-        self.assertEqual(response2.status_code, status.HTTP_201_CREATED)
-
-    def test_planting_plan_area_validation_overlapping_plans_on_different_beds_allowed(self):
-        """Test API allows overlapping plans on different beds."""
-        other_bed = Bed.objects.create(
-            name='Second Test Bed',
-            field=self.field,
-            area_sqm=20.0,
-            project=self.project,
-        )
-
-        response1 = self.client.post('/openfarmplanner/api/planting-plans/', {
-            'culture': self.culture.id,
-            'bed': self.bed.id,
-            'planting_date': '2024-03-01',
-            'area_usage_sqm': 15.0,
-        })
-        self.assertEqual(response1.status_code, status.HTTP_201_CREATED)
-
-        response2 = self.client.post('/openfarmplanner/api/planting-plans/', {
-            'culture': self.culture.id,
-            'bed': other_bed.id,
-            'planting_date': '2024-03-01',
-            'area_usage_sqm': 20.0,
-        })
-        self.assertEqual(response2.status_code, status.HTTP_201_CREATED)
-
-    def test_planting_plan_area_validation_update_excludes_current_plan(self):
-        """Test API update does not count current plan twice."""
-        create_response = self.client.post('/openfarmplanner/api/planting-plans/', {
-            'culture': self.culture.id,
-            'bed': self.bed.id,
-            'planting_date': '2024-03-01',
-            'area_usage_sqm': 10.0,
-        })
-        self.assertEqual(create_response.status_code, status.HTTP_201_CREATED)
-        plan_id = create_response.data['id']
-
-        update_response = self.client.patch(
-            f'/openfarmplanner/api/planting-plans/{plan_id}/',
-            {'area_usage_sqm': 12.0},
-            format='json',
-        )
-        self.assertEqual(update_response.status_code, status.HTTP_200_OK)
-        self.assertEqual(float(update_response.data['area_usage_sqm']), 12.0)
-
-    def test_planting_plan_area_validation_update_partial_rejects_overlap_excess(self):
-        """Test partial update allows the new area usage to be saved."""
-        plan_one_response = self.client.post('/openfarmplanner/api/planting-plans/', {
-            'culture': self.culture.id,
-            'bed': self.bed.id,
-            'planting_date': '2024-03-01',
-            'area_usage_sqm': 12.0,
-        })
-        self.assertEqual(plan_one_response.status_code, status.HTTP_201_CREATED)
-
-        plan_two_response = self.client.post('/openfarmplanner/api/planting-plans/', {
-            'culture': self.culture.id,
-            'bed': self.bed.id,
-            'planting_date': '2024-03-03',
-            'area_usage_sqm': 4.0,
-        })
-        self.assertEqual(plan_two_response.status_code, status.HTTP_201_CREATED)
-
-        update_response = self.client.patch(
-            f"/openfarmplanner/api/planting-plans/{plan_two_response.data['id']}/",
-            {'area_usage_sqm': 9.0},
-            format='json',
-        )
-        self.assertEqual(update_response.status_code, status.HTTP_200_OK)
-        self.assertEqual(float(update_response.data['area_usage_sqm']), 9.0)
-
-    def test_planting_plan_area_validation_boundary_equal_bed_area_allowed(self):
-        """Test API allows overlapping plans whose summed area equals bed capacity."""
-        response1 = self.client.post('/openfarmplanner/api/planting-plans/', {
-            'culture': self.culture.id,
-            'bed': self.bed.id,
-            'planting_date': '2024-03-01',
-            'area_usage_sqm': 12.0,
-        })
-        self.assertEqual(response1.status_code, status.HTTP_201_CREATED)
-
-        response2 = self.client.post('/openfarmplanner/api/planting-plans/', {
-            'culture': self.culture.id,
-            'bed': self.bed.id,
-            'planting_date': '2024-03-03',
-            'area_usage_sqm': 8.0,
-        })
-        self.assertEqual(response2.status_code, status.HTTP_201_CREATED)
-
 
 
 class CultureImportAPITest(DRFAPITestCase):
@@ -1346,548 +812,6 @@ class CultureImportAPITest(DRFAPITestCase):
         self.assertEqual(response.data['created_count'], 1)
         self.assertEqual(response.data['updated_count'], 1)
         self.assertEqual(response.data['skipped_count'], 0)
-
-
-class PlantingPlanAreaInputTest(DRFAPITestCase):
-    """Test area input as m² or plants for PlantingPlan."""
-
-    def setUp(self):
-        """Create test data."""
-        self.user = User.objects.create_user(username='ppuser', email='pp@example.com', password='testpass', is_active=True)
-        self.project = Project.objects.create(name='PP Project', slug='pp-project')
-        ProjectMembership.objects.create(user=self.user, project=self.project, role='admin')
-        self.client.force_authenticate(user=self.user)
-        self.client.defaults['HTTP_X_PROJECT_ID'] = str(self.project.id)
-        # Create location, field, and bed
-        self.location = Location.objects.create(name="Test Farm", project=self.project)
-        self.field = Field.objects.create(
-            name="Test Field",
-            location=self.location,
-            area_sqm=100.00,
-            project=self.project,
-        )
-        self.bed = Bed.objects.create(
-            name="Test Bed",
-            field=self.field,
-            area_sqm=10.00,
-            project=self.project,
-        )
-
-        # Create culture with spacing data
-        self.culture_with_spacing = Culture.objects.create(
-            name="Tomato",
-            growth_duration_days=60,
-            harvest_duration_days=30,
-            row_spacing_m=0.50,  # 50 cm
-            distance_within_row_m=0.40,  # 40 cm
-            project=self.project,
-        )
-
-        # Create culture without spacing data
-        self.culture_no_spacing = Culture.objects.create(
-            name="Cucumber",
-            growth_duration_days=50,
-            harvest_duration_days=20,
-            project=self.project,
-        )
-    
-    def test_plants_per_m2_calculation(self):
-        """Test that plants_per_m2 is calculated correctly."""
-        # 10000 / (50 * 40) = 10000 / 2000 = 5.0
-        expected = 5.0
-        self.assertAlmostEqual(float(self.culture_with_spacing.plants_per_m2), expected, places=2)
-    
-    def test_plants_per_m2_returns_none_when_spacing_missing(self):
-        """Test that plants_per_m2 returns None when spacing is missing."""
-        self.assertIsNone(self.culture_no_spacing.plants_per_m2)
-    
-    def test_area_input_m2_creates_planting_plan(self):
-        """Test creating planting plan with M2 input."""
-        data = {
-            'culture': self.culture_with_spacing.id,
-            'bed': self.bed.id,
-            'planting_date': '2024-03-01',
-            'area_input_value': '2.50',
-            'area_input_unit': 'M2'
-        }
-        
-        response = self.client.post('/openfarmplanner/api/planting-plans/', data, format='json')
-        
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        self.assertEqual(float(response.data['area_usage_sqm']), 2.50)
-    
-    def test_area_input_plants_converts_correctly(self):
-        """Test creating planting plan with PLANTS input converts to m²."""
-        # 10 plants / 5 plants_per_m2 = 2.0 m²
-        data = {
-            'culture': self.culture_with_spacing.id,
-            'bed': self.bed.id,
-            'planting_date': '2024-03-01',
-            'area_input_value': '10',
-            'area_input_unit': 'PLANTS'
-        }
-        
-        response = self.client.post('/openfarmplanner/api/planting-plans/', data, format='json')
-        
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        # 10 / 5 = 2.0
-        self.assertAlmostEqual(float(response.data['area_usage_sqm']), 2.0, places=2)
-    
-    def test_area_input_plants_fails_when_culture_missing(self):
-        """Test that PLANTS input fails when culture is not provided."""
-        data = {
-            'bed': self.bed.id,
-            'planting_date': '2024-03-01',
-            'area_input_value': '10',
-            'area_input_unit': 'PLANTS'
-        }
-        
-        response = self.client.post('/openfarmplanner/api/planting-plans/', data, format='json')
-        
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertIn('culture', response.data)
-    
-    def test_area_input_plants_fails_when_spacing_missing(self):
-        """Test that PLANTS input fails when culture spacing is missing."""
-        data = {
-            'culture': self.culture_no_spacing.id,
-            'bed': self.bed.id,
-            'planting_date': '2024-03-01',
-            'area_input_value': '10',
-            'area_input_unit': 'PLANTS'
-        }
-        
-        response = self.client.post('/openfarmplanner/api/planting-plans/', data, format='json')
-        
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertIn('area_input_unit', response.data)
-    
-    def test_area_input_value_must_be_positive(self):
-        """Test that area_input_value must be greater than 0."""
-        data = {
-            'culture': self.culture_with_spacing.id,
-            'bed': self.bed.id,
-            'planting_date': '2024-03-01',
-            'area_input_value': '0',
-            'area_input_unit': 'M2'
-        }
-        
-        response = self.client.post('/openfarmplanner/api/planting-plans/', data, format='json')
-        
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertIn('area_input_value', response.data)
-    
-    def test_area_input_unit_required_when_value_provided(self):
-        """Test that area_input_unit is required when area_input_value is provided."""
-        data = {
-            'culture': self.culture_with_spacing.id,
-            'bed': self.bed.id,
-            'planting_date': '2024-03-01',
-            'area_input_value': '2.50'
-            # Missing area_input_unit
-        }
-        
-        response = self.client.post('/openfarmplanner/api/planting-plans/', data, format='json')
-        
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertIn('area_input_unit', response.data)
-    
-    def test_area_input_plants_update_existing_plan(self):
-        """Test updating existing planting plan with PLANTS input."""
-        # Create initial plan
-        plan = PlantingPlan.objects.create(
-            culture=self.culture_with_spacing,
-            bed=self.bed,
-            planting_date=date(2024, 3, 1),
-            area_usage_sqm=1.0,
-            project=self.project,
-        )
-        
-        # Update with PLANTS input: 15 plants / 5 plants_per_m2 = 3.0 m²
-        data = {
-            'culture': self.culture_with_spacing.id,
-            'bed': self.bed.id,
-            'planting_date': '2024-03-01',
-            'area_input_value': '15',
-            'area_input_unit': 'PLANTS'
-        }
-        
-        response = self.client.put(f'/openfarmplanner/api/planting-plans/{plan.id}/', data, format='json')
-        
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertAlmostEqual(float(response.data['area_usage_sqm']), 3.0, places=2)
-
-    def test_area_input_m2_put_update_does_not_return_server_error(self):
-        """Test full PUT update with M2 input does not crash and updates area."""
-        plan = PlantingPlan.objects.create(
-            culture=self.culture_with_spacing,
-            bed=self.bed,
-            planting_date=date(2026, 1, 1),
-            area_usage_sqm=1.0,
-            project=self.project,
-            cultivation_type='pre_cultivation',
-        )
-
-        response = self.client.put(
-            f'/openfarmplanner/api/planting-plans/{plan.id}/',
-            {
-                'culture': self.culture_with_spacing.id,
-                'bed': self.bed.id,
-                'planting_date': '2026-01-01',
-                'quantity': None,
-                'notes': '',
-                'cultivation_type': 'pre_cultivation',
-                'area_input_value': 3,
-                'area_input_unit': 'M2',
-            },
-            format='json',
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(float(response.data['area_usage_sqm']), 3.0)
-
-
-class NoteAttachmentApiTest(DRFAPITestCase):
-    def setUp(self):
-        self.user = User.objects.create_user(username='attachapiuser', email='attachapi@example.com', password='testpass', is_active=True)
-        self.project = Project.objects.create(name='Attachment API Project', slug='attachment-api-project')
-        ProjectMembership.objects.create(user=self.user, project=self.project, role='admin')
-        self.client.force_authenticate(user=self.user)
-        self.client.defaults['HTTP_X_PROJECT_ID'] = str(self.project.id)
-        self.location = Location.objects.create(name="Attachment Location", project=self.project)
-        self.field = Field.objects.create(name="Attachment Field", location=self.location, project=self.project)
-        self.bed = Bed.objects.create(name="Attachment Bed", field=self.field, project=self.project)
-        self.culture = Culture.objects.create(
-            name="Attachment Culture",
-            growth_duration_days=7,
-            harvest_duration_days=2,
-            project=self.project,
-        )
-        self.plan = PlantingPlan.objects.create(
-            culture=self.culture,
-            bed=self.bed,
-            planting_date=date(2024, 3, 1),
-            project=self.project,
-        )
-
-    @patch('farm.notes.views.process_note_image')
-    def test_upload_list_delete_attachment(self, mock_process):
-        mock_process.return_value = (
-            SimpleUploadedFile('processed.webp', b'processed', content_type='image/webp'),
-            {
-                'width': 1280,
-                'height': 720,
-                'size_bytes': 9,
-                'mime_type': 'image/webp',
-            },
-        )
-        upload = SimpleUploadedFile('raw.jpg', b'raw', content_type='image/jpeg')
-
-        upload_response = self.client.post(
-            f'/openfarmplanner/api/notes/{self.plan.id}/attachments/',
-            {'image': upload},
-            format='multipart',
-        )
-        self.assertEqual(upload_response.status_code, status.HTTP_201_CREATED)
-        self.assertLessEqual(upload_response.data['width'], 1280)
-
-        list_response = self.client.get(f'/openfarmplanner/api/notes/{self.plan.id}/attachments/')
-        self.assertEqual(list_response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(list_response.data), 1)
-
-        attachment_id = upload_response.data['id']
-        delete_response = self.client.delete(f'/openfarmplanner/api/attachments/{attachment_id}/')
-        self.assertEqual(delete_response.status_code, status.HTTP_204_NO_CONTENT)
-
-
-    @patch(
-        'farm.notes.views.process_note_image',
-        side_effect=__import__('farm.image_processing', fromlist=['ImageProcessingBackendUnavailableError']).ImageProcessingBackendUnavailableError('Image processing backend is not available. Install Pillow in the backend environment.'),
-    )
-    def test_attachment_upload_returns_503_when_processing_backend_missing(self, _mock_process):
-        upload = SimpleUploadedFile('raw.jpg', b'raw', content_type='image/jpeg')
-        response = self.client.post(
-            f'/openfarmplanner/api/notes/{self.plan.id}/attachments/',
-            {'image': upload},
-            format='multipart',
-        )
-        self.assertEqual(response.status_code, status.HTTP_503_SERVICE_UNAVAILABLE)
-
-    @patch(
-        'farm.notes.views.process_note_image',
-        side_effect=__import__('farm.image_processing', fromlist=['ImageProcessingError']).ImageProcessingError('bad image'),
-    )
-    def test_invalid_attachment_upload_returns_400(self, _mock_process):
-        upload = SimpleUploadedFile('not-image.txt', b'text', content_type='text/plain')
-        response = self.client.post(
-            f'/openfarmplanner/api/notes/{self.plan.id}/attachments/',
-            {'image': upload},
-            format='multipart',
-        )
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-
-    def test_list_attachments_for_other_project_is_forbidden(self):
-        other_user = User.objects.create_user(username='attachother', email='attachother@example.com', password='testpass', is_active=True)
-        other_project = Project.objects.create(name='Other Attachment Project', slug='other-attachment-project')
-        ProjectMembership.objects.create(user=other_user, project=other_project, role='admin')
-
-        self.client.force_authenticate(user=other_user)
-        self.client.defaults['HTTP_X_PROJECT_ID'] = str(other_project.id)
-
-        response = self.client.get(f'/openfarmplanner/api/notes/{self.plan.id}/attachments/')
-        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
-
-
-class PlantingPlanAttachmentCountApiTest(DRFAPITestCase):
-    def setUp(self):
-        self.user = User.objects.create_user(username='attachcountuser', email='attachcount@example.com', password='testpass', is_active=True)
-        self.project = Project.objects.create(name='Attach Count Project', slug='attach-count-project')
-        ProjectMembership.objects.create(user=self.user, project=self.project, role='admin')
-        self.client.force_authenticate(user=self.user)
-        self.client.defaults['HTTP_X_PROJECT_ID'] = str(self.project.id)
-        self.location = Location.objects.create(name="Plan Location", project=self.project)
-        self.field = Field.objects.create(name="Plan Field", location=self.location, project=self.project)
-        self.bed = Bed.objects.create(name="Plan Bed", field=self.field, project=self.project)
-        self.culture = Culture.objects.create(name="Plan Culture", growth_duration_days=7, harvest_duration_days=2, project=self.project)
-
-        self.plan_without_attachments = PlantingPlan.objects.create(
-            culture=self.culture, bed=self.bed, planting_date=date(2024, 3, 1), notes='No attachments',
-            project=self.project,
-        )
-        self.plan_with_attachments = PlantingPlan.objects.create(
-            culture=self.culture, bed=self.bed, planting_date=date(2024, 3, 2), notes='With attachments',
-            project=self.project,
-        )
-
-        NoteAttachment.objects.create(
-            planting_plan=self.plan_with_attachments,
-            image='notes/test-1.webp',
-            mime_type='image/webp',
-            width=100,
-            height=100,
-            size_bytes=1000,
-            project=self.project,
-        )
-        NoteAttachment.objects.create(
-            planting_plan=self.plan_with_attachments,
-            image='notes/test-2.webp',
-            mime_type='image/webp',
-            width=100,
-            height=100,
-            size_bytes=1000,
-            project=self.project,
-        )
-
-    def test_planting_plan_list_contains_attachment_count(self):
-        response = self.client.get('/openfarmplanner/api/planting-plans/')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        by_id = {item['id']: item for item in response.data['results']}
-        self.assertEqual(by_id[self.plan_without_attachments.id]['note_attachment_count'], 0)
-        self.assertEqual(by_id[self.plan_with_attachments.id]['note_attachment_count'], 2)
-
-    def test_planting_plan_list_query_count_stays_stable(self):
-        # 5 queries: 1 SAVEPOINT + 1 project membership lookup + 1 COUNT (pagination)
-        # + 1 SELECT with annotation + 1 RELEASE SAVEPOINT (ATOMIC_REQUESTS wraps each
-        # request in its own transaction, nested as a savepoint under the test's outer
-        # transaction).
-        with self.assertNumQueries(5):
-            response = self.client.get('/openfarmplanner/api/planting-plans/')
-            self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-
-class PlantingPlanRemainingAreaApiTest(DRFAPITestCase):
-    """Tests for remaining-area endpoint on planting plans."""
-
-    def setUp(self):
-        self.user = User.objects.create_user(username='remaininguser', email='remaining@example.com', password='testpass', is_active=True)
-        self.project = Project.objects.create(name='Remaining Area Project', slug='remaining-area-project')
-        ProjectMembership.objects.create(user=self.user, project=self.project, role='admin')
-        self.client.force_authenticate(user=self.user)
-        self.client.defaults['HTTP_X_PROJECT_ID'] = str(self.project.id)
-        self.location = Location.objects.create(name='Remaining Location', project=self.project)
-        self.field = Field.objects.create(name='Remaining Field', location=self.location, project=self.project)
-        self.bed = Bed.objects.create(name='Remaining Bed', field=self.field, area_sqm=20, project=self.project)
-        self.other_bed = Bed.objects.create(name='Other Bed', field=self.field, area_sqm=15, project=self.project)
-        self.culture = Culture.objects.create(
-            name='Lettuce',
-            growth_duration_days=30,
-            harvest_duration_days=10,
-            project=self.project,
-        )
-
-        self.plan_one = PlantingPlan.objects.create(
-            culture=self.culture,
-            bed=self.bed,
-            planting_date=date(2024, 3, 1),
-            area_usage_sqm=6,
-            project=self.project,
-        )
-        self.plan_two = PlantingPlan.objects.create(
-            culture=self.culture,
-            bed=self.bed,
-            planting_date=date(2024, 3, 15),
-            area_usage_sqm=4,
-            project=self.project,
-        )
-        PlantingPlan.objects.create(
-            culture=self.culture,
-            bed=self.other_bed,
-            planting_date=date(2024, 3, 10),
-            area_usage_sqm=8,
-            project=self.project,
-        )
-
-    def test_remaining_area_returns_overlap_sum(self):
-        response = self.client.get(
-            '/openfarmplanner/api/planting-plans/remaining-area/',
-            {
-                'bed_id': self.bed.id,
-                'start_date': '2024-03-20',
-                'end_date': '2024-04-10',
-            },
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data['bed_id'], self.bed.id)
-        self.assertEqual(response.data['bed_area_sqm'], 20.0)
-        self.assertEqual(response.data['overlapping_used_area_sqm'], 10.0)
-        self.assertEqual(response.data['remaining_area_sqm'], 10.0)
-
-    def test_remaining_area_excludes_current_plan(self):
-        response = self.client.get(
-            '/openfarmplanner/api/planting-plans/remaining-area/',
-            {
-                'bed_id': self.bed.id,
-                'start_date': '2024-03-20',
-                'end_date': '2024-04-10',
-                'exclude_plan_id': self.plan_two.id,
-            },
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data['overlapping_used_area_sqm'], 6.0)
-        self.assertEqual(response.data['remaining_area_sqm'], 14.0)
-
-    def test_remaining_area_validates_required_params(self):
-        response = self.client.get('/openfarmplanner/api/planting-plans/remaining-area/')
-
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertIn('detail', response.data)
-
-    def test_remaining_area_rejects_invalid_date_range(self):
-        response = self.client.get(
-            '/openfarmplanner/api/planting-plans/remaining-area/',
-            {
-                'bed_id': self.bed.id,
-                'start_date': '2024-04-10',
-                'end_date': '2024-03-20',
-            },
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertIn('detail', response.data)
-
-    def test_remaining_area_rejects_invalid_bed_id_type(self):
-        response = self.client.get(
-            '/openfarmplanner/api/planting-plans/remaining-area/',
-            {
-                'bed_id': 'abc',
-                'start_date': '2024-03-20',
-                'end_date': '2024-04-10',
-            },
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertIn('detail', response.data)
-
-    def test_remaining_area_rejects_bed_from_other_project(self):
-        other_project = Project.objects.create(name='Other Remaining Project', slug='other-remaining-project')
-        other_location = Location.objects.create(name='Other Remaining Location', project=other_project)
-        other_field = Field.objects.create(name='Other Remaining Field', location=other_location, project=other_project)
-        other_bed = Bed.objects.create(name='Other Remaining Bed', field=other_field, area_sqm=12, project=other_project)
-
-        response = self.client.get(
-            '/openfarmplanner/api/planting-plans/remaining-area/',
-            {
-                'bed_id': other_bed.id,
-                'start_date': '2024-03-20',
-                'end_date': '2024-04-10',
-            },
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
-
-
-class CultureLayoutApiTest(DRFAPITestCase):
-    """Tests for the bed/field layout endpoint on locations."""
-
-    def setUp(self):
-        self.user = User.objects.create_user(username='layoutuser', email='layoutuser@example.com', password='testpass', is_active=True)
-        self.project = Project.objects.create(name='Layout Project', slug='layout-project')
-        ProjectMembership.objects.create(user=self.user, project=self.project, role='admin')
-        self.client.force_authenticate(user=self.user)
-        self.client.defaults['HTTP_X_PROJECT_ID'] = str(self.project.id)
-        self.supplier = Supplier.objects.create(name='Default Supplier', homepage_url='https://supplier.example', project=self.project)
-        self.culture = Culture.objects.create(name='Tomate', variety='Roma', supplier=self.supplier, supplier_product_url='https://supplier.example/tomate', project=self.project)
-
-    def test_layouts_get_and_put(self):
-        location = Location.objects.create(name='Layout test location', project=self.project)
-        field = Field.objects.create(name='Layout test field', location=location, project=self.project)
-        bed = Bed.objects.create(name='Layout test bed', field=field, area_sqm=5, project=self.project)
-
-        payload = {
-            'bed_layouts': [
-                {'bed': bed.id, 'location': location.id, 'x': 33.5, 'y': 44.5, 'version': 1},
-            ],
-            'field_layouts': [
-                {'field': field.id, 'location': location.id, 'x': 66.0, 'y': 88.0, 'version': 1},
-            ],
-        }
-        put_response = self.client.put(
-            f'/openfarmplanner/api/locations/{location.id}/layouts/',
-            payload,
-            format='json',
-        )
-        self.assertEqual(put_response.status_code, status.HTTP_200_OK)
-        self.assertEqual(put_response.data['bed_layouts'][0]['bed'], bed.id)
-        self.assertEqual(put_response.data['field_layouts'][0]['field'], field.id)
-
-        get_response = self.client.get(f'/openfarmplanner/api/locations/{location.id}/layouts/')
-        self.assertEqual(get_response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(get_response.data['bed_layouts']), 1)
-        self.assertEqual(get_response.data['bed_layouts'][0]['x'], 33.5)
-        self.assertEqual(len(get_response.data['field_layouts']), 1)
-        self.assertEqual(get_response.data['field_layouts'][0]['x'], 66.0)
-
-    def test_layouts_reject_bed_from_other_location(self):
-        location = Location.objects.create(name='Layout test source location', project=self.project)
-        other_location = Location.objects.create(name='Secondary location', project=self.project)
-        other_field = Field.objects.create(name='Secondary field', location=other_location, project=self.project)
-        other_bed = Bed.objects.create(name='Secondary bed', field=other_field, area_sqm=5, project=self.project)
-
-        response = self.client.put(
-            f'/openfarmplanner/api/locations/{location.id}/layouts/',
-            {'bed_layouts': [{'bed': other_bed.id, 'location': location.id, 'x': 1, 'y': 1}], 'field_layouts': []},
-            format='json',
-        )
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertIn('does not belong to location', response.data['detail'])
-        self.assertFalse(BedLayout.objects.filter(bed=other_bed).exists())
-
-    def test_layouts_reject_field_from_other_location(self):
-        location = Location.objects.create(name='Layout test source location', project=self.project)
-        other_location = Location.objects.create(name='Secondary location', project=self.project)
-        other_field = Field.objects.create(name='Secondary field', location=other_location, project=self.project)
-
-        response = self.client.put(
-            f'/openfarmplanner/api/locations/{location.id}/layouts/',
-            {'bed_layouts': [], 'field_layouts': [{'field': other_field.id, 'location': location.id, 'x': 1, 'y': 1}]},
-            format='json',
-        )
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertIn('does not belong to location', response.data['detail'])
-        self.assertFalse(FieldLayout.objects.filter(field=other_field).exists())
 
 
 class PublicCultureLibraryApiTest(DRFAPITestCase):
@@ -2200,3 +1124,4 @@ class PublicCultureLibraryApiTest(DRFAPITestCase):
         response = self.client.post(f'/openfarmplanner/api/public-cultures/{public_culture.id}/import/', {}, format='json')
 
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
